+++ src/js/videojs.record.js
@@ -1620,13 +1620,70 @@ class Record extends Plugin {
     }
 
     /**
+     * Change the video input device.
+     *
+     * @param {string} deviceId - Id of the video input device.
+     */
+    setVideoInput(deviceId) {
+        if (this.recordVideo === Object(this.recordVideo)) {
+            // already using video constraints
+            this.recordVideo.deviceId = {exact: deviceId};
+
+        } else if (this.recordVideo === true) {
+            // not using video constraints already, so force it
+            this.recordVideo = {
+                deviceId: {exact: deviceId}
+            };
+        }
+
+        // release existing device
+        this.stopDevice();
+
+        // ask for video input device permissions and start device
+        this.getDevice();
+    }
+
+    /**
+     * Change the audio input device.
+     *
+     * @param {string} deviceId - Id of the audio input device.
+     */
+    setAudioInput(deviceId) {
+        if (this.recordAudio === Object(this.recordAudio)) {
+            // already using audio constraints
+            this.recordAudio.deviceId = {exact: deviceId};
+
+        } else if (this.recordAudio === true) {
+            // not using audio constraints already, so force it
+            this.recordAudio = {
+                deviceId: {exact: deviceId}
+            };
+        }
+
+        // update wavesurfer microphone plugin constraints
+        switch (this.getRecordType()) {
+            case AUDIO_ONLY:
+                this.surfer.surfer.microphone.constraints = {
+                    video: false,
+                    audio: this.recordAudio
+                };
+                break;
+        }
+
+        // release existing device
+        this.stopDevice();
+
+        // ask for audio input device permissions and start device
+        this.getDevice();
+    }
+
+    /**
      * Change the audio output device.
      *
      * @param {string} deviceId - Id of audio output device.
      */
     setAudioOutput(deviceId) {
         let errorMessage;
-
         switch (this.getRecordType()) {
             case AUDIO_ONLY:
                 // use wavesurfer
@@ -1654,7 +1711,7 @@ class Record extends Plugin {
                         errorMessage = 'Browser does not support audio output device selection.';
                     }
                 } else {
-                    errorMessage = 'Invalid deviceId: ' + deviceId;
+                    errorMessage = `Invalid deviceId: ${deviceId}`;
                 }
                 break;
         }